The purpose for this application is to count lines in a text file.
Once it is submitted, it will examine the file line by line and
determine whether the lines are valid (meaning no blank lines or
comments). It will count the total number of "real lines" in the file,
and display it for the user to see. It will include documentation and
notes to better understand the process.